Just in case, allow fractional -framecap

+++ Descent3/init.cpp
@@ -315,7 +315,7 @@ void PreInitD3Systems()
 	iframelmtarg = FindArg("-framecap");
 	if(iframelmtarg)
 	{
-		Min_allowed_frametime = (1.0/atoi(GameArgs[iframelmtarg+1]));
+		Min_allowed_frametime = (1.0/atof(GameArgs[iframelmtarg+1]));
 		mprintf((0,"Using %f as a minimum frametime\n",Min_allowed_frametime));
 	}
 	else
